Add temporary hack to save HTML contents to string
This change doesn’t need review. It’s just a temporary hack for use on
this branch to test piping of output to an external process and getting
a result back.

For syntax highlighting, we’ll need to end up piping JSON out, not HTML
— but since we don’t yet have the actual external tool we need for
consuming the JSON, this enables us to instead pipe out HTML for now
(and to just run it through an external process that echoes it back).