Emit each HTML output filename as it gets written
This change causes wattsi to emit messages with the name of each HTML output
filename as it gets written.

The syntax-highlighting mechanism increases the need for more messages from
wattsi as it runs — because it causes wattsi to take longer to run; in
particular, it takes longer while it’s writing the HTML output files.

So, without some kind of progress indication getting output while wattsi is
writing the files, it looks like it’s hanging. Emitting the filenames provides
the needed progress indication.